+++ src/main/java/com/epam/provider/service/ProfileService.java
@@ -18,7 +18,7 @@ public interface ProfileService {
    * @param amount money to deposit
    * @param profileId id of client who pays
    */
-  void addBalance(Integer amount, Integer profileId) throws ServiceException;
+  void addBalance(Double amount, Integer profileId) throws ServiceException;
 
   /**
    * Method selects user from database by login and pass. If no user was found the empty object is
